+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 
 ### 🐞 Fixes
 
+- [#874](https://github.com/estruyf/vscode-front-matter/issues/874): Fix media snippet markup insertion to article content's
+
 ## [10.5.0] - 2024-10-21 - [Release notes](https://beta.frontmatter.codes/updates/v10.5.0)
 
 ### 🎨 Enhancements